Create Environment_setup.py
Creates a grid_world environment defining its obstacles, locations and rewards. 
It also allows to draw the corresponding state value functions and policies over the grid disposition.
It has been based in the RL lab code supplied by the GTA's and the Dr. Aldo Faisal in the RL course of the Imperial College of London.